bpf: Move linux specific code into _linux.go files
[ upstream commit 1d28c269b6c54de1b442a044a9bc76accb0e9058 ]
